Index game containers with signed types
RA's TFixedIHeapClass::Ptr and Raw_Ptr take int, as TD's already did, and TD's
VectorClass capacity is base::ssize like RA's, with the same bounds DCHECK on
subscripts. Heap free flags, cell redraw flags, vector allocation and the index
table's qsort/bsearch cross into size_t through base::ToSize. Removes 319
sign-conversion reports across both games.

+++ src/ra/serialize.cc
@@ -39,7 +39,7 @@ ObjectClass* Slot(TFixedIHeapClass<T>& heap, int index, ArchiveReader& ar) {
     ar.Fail("saved object index outside its heap");
     return nullptr;
   }
-  return heap.Raw_Ptr(static_cast<std::size_t>(index));
+  return heap.Raw_Ptr(index);
 }
 
 bool IsTechnoKind(RTTIType kind) {
